:triangular_flag_on_post: Checkpoint
This commit has discovered a bug with Meteor throwing an error saying a certain academic term is not a valid Academic Term. Even though if you search the id of that Academic Term in the database, it is there.
I traced it from the function AcademicTerms.getShortName(). This error appears when you are logged in as a certain user
and you try to view an opportunity/course page that has a review attached to it by that certain user.