Found a work around for the reloading problem.
In the StudentProtectedRoute, we are looking to see if Meteor.user() is defined. If it is not we redirect to the LandingPage and set the state to the current page.
The LandingPage checks the state to see if there is a location to go to and redirects back to the page.